dissatisfying, slow solutions for day 16
At first, I tried to iterate a billion times. that takes too long. then,
I tried to use a hash to cache the results of an iteration on any
particular input. that also takes... a very long time. I didn't wait
long enough to find out how long. Then, I looked at discussion on the
internet and realized that I should have been looking for a cycle in the
inputs and outputs, so that I could skip forward to the end to get the
right answer in a reasonable amount of time. But even looking for the
first cycle and returning the answer after that is so slow that I
haven't finished running it yet. Ugh.